refactor: move tests to integrationTest
