+++ buildSrc/src/testing/kotlin/wolkenschloss/gradle/docker/Extensions.kt
@@ -1,12 +1,9 @@
-package wolkenschloss.gradle.docker.testing
+package wolkenschloss.gradle.docker
 
 import com.github.dockerjava.api.model.Image
 import org.gradle.api.file.FileSystemLocation
 import org.gradle.api.file.FileSystemLocationProperty
 import org.gradle.api.tasks.TaskProvider
-import org.gradle.testkit.runner.BuildResult
-import org.gradle.testkit.runner.GradleRunner
-import wolkenschloss.gradle.docker.BuildImageTask
 import java.io.File
 
 val Image.shortId: String
